+++ app/src/main/java/project_645/BufferManager.java
@@ -17,7 +17,7 @@ public abstract class BufferManager {
      */
 
 
-    public abstract Page getPage(int pageId, File dataFile) throws Exception;
+    public abstract Page getPage(long pageId, File dataFile) throws Exception;
 
     /**
          * Creates a new page.
@@ -31,13 +31,13 @@ public abstract class BufferManager {
          * @param pageId The ID of the page to mark as dirty.
          */
 
-    public abstract void markDirty(int pageId, File dataFile);
+    public abstract void markDirty(long pageId, File dataFile);
 
     /**
          * Unpins a page in the buffer pool, allowing it to be evicted if necessary.
          * @param pageId The ID of the page to unpin.
          */
-    public abstract void unpinPage(int pageId, File dataFile);
+    public abstract void unpinPage(long pageId, File dataFile);
 
     /**
      *
